Some versioning tweaks so this builds locally

+++ ml-agents/mlagents/trainers/buffer.py
@@ -207,7 +207,7 @@ class AgentBufferField(list):
         self[:] = []
 
     def padded_to_batch(
-        self, pad_value: np.float = 0, dtype: np.dtype = np.float32
+        self, pad_value: float = 0, dtype: np.dtype = np.float32
     ) -> Union[np.ndarray, List[np.ndarray]]:
         """
         Converts this AgentBufferField (which is a List[BufferEntry]) into a numpy array
